refactor(2-2): 패키지 이동 및 클래스 이름 변경
