+++ src/main/java/nextstep/security/oauth2/OAuth2AuthenticationException.java
@@ -1,0 +1,8 @@
+package nextstep.security.oauth2;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNAUTHORIZED)
+public class OAuth2AuthenticationException extends RuntimeException {
+}
